refactor: allow RegistrationEndpoint to take several claims classes
